pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "stmtools"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "dask[complete]",
    "xarray",
    "numpy",
    "rasterio",
    "geopandas",
    "affine",
    "zarr",
    "distributed",
]
description = "space-time matrix for PS-InSAR application"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Ou Ku", email = "o.ku@esciencecenter.nl"},
    {name = "Fakhereh Alidoost"},
    {name = "Pranav Chandramouli"},
    {name = "Meiert Willem Grootes"},
    {name = "Francesco Nattino"},
    {name = "Thijs van Lankveld"},
    {name = "Freek van Leijen"},
]
keywords = ["stm", "sar", "insar", "space-time", "psi"]
classifiers=[
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]

[project.urls]
repository = "https://github.com/MotionbyLearning/stmtools"
doi = "https://doi.org/10.5281/zenodo.7717112"
documentation = "https://motionbylearning.github.io/stmtools/"
changelog = "https://motionbylearning.github.io/stmtools/CHANGELOG/"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pycodestyle",
    "ruff",
]
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocs-jupyter",
    "mkdocstrings[python]",
    "mkdocs-gen-files",
]
demo = [
    "jupyterlab",
    "matplotlib",
]

[tool.setuptools]
packages = ["stmtools"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov --cov-report term"
testpaths = [
    "tests",
]

[tool.coverage.run]
branch = true
source = ["stmtools"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 99
indent-width = 4

# Assume Python 3.10
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "B",  # flake8-bugbear
  "D",  # pydocstyle
  "I",  # isort
  "N",  # PEP8-naming
  "UP",  # pyupgrade (upgrade syntax to current syntax)
  "PLE",  # Pylint error https://github.com/charliermarsh/ruff#error-ple
]

ignore = ["D100", "D101", "D104", "D105", "D106", "D107", "D203", "D213"] # docstring style

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.per-file-ignores]
"tests/**" = ["D"]